Extract out build command
